Make Markdown linguist-detectable
Before this change, repository gets detected as:

Shell 100.0%

After this change, repository gets detected as:

Markdown 99.9%
Shell 0.1%

This change was added for a cosmetic effect on GitHub. It might be
good to go away once this repository is no longer hosted at GitHub, or
if GitHub ceases to use the line, or GitHub ceases to exist.